src/equity-eye/scraper/article_collector/requirements.txt
newsapi-python
requests
beautifulsoup4
spacy
fuzzywuzzy
python-Levenshtein